Move the stress session into its own file
netplay.c had grown to 1219 lines. The session, its menu driving and its input
generation move to netplay_stress.c, with netplay_base.h holding the State the
two share and the hooks netplay.c still implements.
