+++ src/main.c
@@ -2,6 +2,7 @@
 #include "common.h"
 #if NETPLAY_ENABLED
 #include "platform/netplay/netplay.h"
+#include "platform/netplay/netplay_stress.h"
 #endif
 #include "sf33rd/AcrSDK/common/mlPAD.h"
 #include "sf33rd/AcrSDK/ps2/flps2debug.h"
@@ -263,7 +264,7 @@ void Main_StepFrame() {
     keyConvert();
 
 #if NETPLAY_ENABLED
-    Netplay_InjectStressBootInput();
+    Stress_InjectBootInput();
 #endif
 
 #if DEBUG
@@ -302,7 +303,7 @@ void Main_StepFrame() {
         njdp2d_draw();
         Netplay_TickMatchmaking();
         Netplay_TickDirectP2P();
-        Netplay_TickStress();
+        Stress_Tick();
     }
 #else
     njUserMain();
