+++ src/platform/netplay/netplay.h
@@ -23,10 +23,6 @@ void Netplay_SetParams(int player, const char* ip);
 bool Netplay_IsDirectP2PConfigured(); // true when a remote was given on the command line
 void Netplay_BeginDirectP2P();
 void Netplay_TickDirectP2P();
-void Netplay_SetStressOutputDir(const char* directory);
-void Netplay_BeginStress(int seed, int check_distance, int frames);
-void Netplay_TickStress();
-void Netplay_InjectStressBootInput();
 void Netplay_SetMatchmakingParams(const char* server_ip, int server_port);
 void Netplay_BeginMatchmaking();
 void Netplay_TickMatchmaking();
